Split `PermissionGrant` types
The PermissionGrants.ts file was getting really big (and the linter was
about to get upset).  This change breaks it across reasonable boundaries
to make it easier to interact with.
