+++ src/types/PermissionGrant.ts
@@ -1,7 +1,18 @@
 import { ajv } from '../ajv';
 import { isEmpty } from '../arrays';
-import { BaseFieldCategory } from './BaseField';
 import { keycloakIdSchema } from './KeycloakId';
+import {
+	permissionGrantConditionSchema,
+	type PermissionGrantCondition,
+} from './PermissionGrantCondition';
+import {
+	contextEntityKeyProperties,
+	isPermissionGrantKeyedContextEntityType,
+	jsonSchemaTypeForPermissionGrantEntityKeyType,
+	PermissionGrantEntityType,
+	type PermissionGrantEntityKeyValueType,
+	type PermissionGrantKeyedContextEntityType,
+} from './PermissionGrantEntityType';
 import {
 	PermissionGrantGranteeType,
 	permissionGrantGranteeTypeSchema,
@@ -18,38 +29,6 @@ import type { User } from './User';
 import type { Writable } from './Writable';
 import type { ValidateFunction } from 'ajv';
 
-enum PermissionGrantEntityType {
-	FUNDER = 'funder',
-	CHANGEMAKER = 'changemaker',
-	DATA_PROVIDER = 'dataProvider',
-	OPPORTUNITY = 'opportunity',
-	PROPOSAL = 'proposal',
-	PROPOSAL_VERSION = 'proposalVersion',
-	APPLICATION_FORM = 'applicationForm',
-	APPLICATION_FORM_FIELD = 'applicationFormField',
-	PROPOSAL_FIELD_VALUE = 'proposalFieldValue',
-	SOURCE = 'source',
-	BULK_UPLOAD = 'bulkUpload',
-	CHANGEMAKER_FIELD_VALUE = 'changemakerFieldValue',
-	ANY = 'any',
-}
-
-enum PermissionGrantEntityKeyType {
-	ID = 'id',
-	SHORT_CODE = 'shortCode',
-}
-
-interface PermissionGrantEntityKeyValueType {
-	[PermissionGrantEntityKeyType.ID]: Id;
-	[PermissionGrantEntityKeyType.SHORT_CODE]: string;
-}
-
-interface PermissionGrantCondition {
-	property: string;
-	operator: string;
-	value: string[];
-}
-
 // The key name for each grantee type
 const granteeKeyProperties = {
 	[PermissionGrantGranteeType.USER]: {
@@ -59,148 +38,6 @@ const granteeKeyProperties = {
 		keyName: 'granteeKeycloakOrganizationId',
 	},
 } as const satisfies Record<PermissionGrantGranteeType, { keyName: string }>;
-
-// The primary key name + data type for each context entity
-const contextEntityKeyProperties = {
-	[PermissionGrantEntityType.CHANGEMAKER]: {
-		keyName: 'changemakerId',
-		keyType: PermissionGrantEntityKeyType.ID,
-	},
-	[PermissionGrantEntityType.FUNDER]: {
-		keyName: 'funderShortCode',
-		keyType: PermissionGrantEntityKeyType.SHORT_CODE,
-	},
-	[PermissionGrantEntityType.DATA_PROVIDER]: {
-		keyName: 'dataProviderShortCode',
-		keyType: PermissionGrantEntityKeyType.SHORT_CODE,
-	},
-	[PermissionGrantEntityType.OPPORTUNITY]: {
-		keyName: 'opportunityId',
-		keyType: PermissionGrantEntityKeyType.ID,
-	},
-	[PermissionGrantEntityType.PROPOSAL]: {
-		keyName: 'proposalId',
-		keyType: PermissionGrantEntityKeyType.ID,
-	},
-	[PermissionGrantEntityType.PROPOSAL_VERSION]: {
-		keyName: 'proposalVersionId',
-		keyType: PermissionGrantEntityKeyType.ID,
-	},
-	[PermissionGrantEntityType.APPLICATION_FORM]: {
-		keyName: 'applicationFormId',
-		keyType: PermissionGrantEntityKeyType.ID,
-	},
-	[PermissionGrantEntityType.APPLICATION_FORM_FIELD]: {
-		keyName: 'applicationFormFieldId',
-		keyType: PermissionGrantEntityKeyType.ID,
-	},
-	[PermissionGrantEntityType.PROPOSAL_FIELD_VALUE]: {
-		keyName: 'proposalFieldValueId',
-		keyType: PermissionGrantEntityKeyType.ID,
-	},
-	[PermissionGrantEntityType.SOURCE]: {
-		keyName: 'sourceId',
-		keyType: PermissionGrantEntityKeyType.ID,
-	},
-	[PermissionGrantEntityType.BULK_UPLOAD]: {
-		keyName: 'bulkUploadTaskId',
-		keyType: PermissionGrantEntityKeyType.ID,
-	},
-	[PermissionGrantEntityType.CHANGEMAKER_FIELD_VALUE]: {
-		keyName: 'changemakerFieldValueId',
-		keyType: PermissionGrantEntityKeyType.ID,
-	},
-} as const satisfies Omit<
-	Record<
-		PermissionGrantEntityType,
-		{ keyName: string; keyType: PermissionGrantEntityKeyType }
-	>,
-	PermissionGrantEntityType.ANY
->;
-
-type PermissionGrantKeyedContextEntityType =
-	keyof typeof contextEntityKeyProperties;
-
-// Scopes that are valid against any context entity type, appended to a
-// context's native scopes by `getScopesForContextEntityType` below.
-const universalScopes = [PermissionGrantEntityType.ANY] as const;
-
-const contextEntityTypeNativeScopes = {
-	[PermissionGrantEntityType.CHANGEMAKER]: [
-		PermissionGrantEntityType.CHANGEMAKER,
-		PermissionGrantEntityType.CHANGEMAKER_FIELD_VALUE,
-		PermissionGrantEntityType.PROPOSAL,
-		PermissionGrantEntityType.PROPOSAL_FIELD_VALUE,
-	],
-	[PermissionGrantEntityType.FUNDER]: [
-		PermissionGrantEntityType.FUNDER,
-		PermissionGrantEntityType.OPPORTUNITY,
-		PermissionGrantEntityType.APPLICATION_FORM,
-		PermissionGrantEntityType.PROPOSAL,
-		PermissionGrantEntityType.PROPOSAL_FIELD_VALUE,
-	],
-	[PermissionGrantEntityType.DATA_PROVIDER]: [
-		PermissionGrantEntityType.DATA_PROVIDER,
-	],
-	[PermissionGrantEntityType.OPPORTUNITY]: [
-		PermissionGrantEntityType.OPPORTUNITY,
-		PermissionGrantEntityType.APPLICATION_FORM,
-		PermissionGrantEntityType.PROPOSAL,
-		PermissionGrantEntityType.PROPOSAL_FIELD_VALUE,
-	],
-	[PermissionGrantEntityType.PROPOSAL]: [
-		PermissionGrantEntityType.PROPOSAL,
-		PermissionGrantEntityType.PROPOSAL_FIELD_VALUE,
-	],
-	[PermissionGrantEntityType.PROPOSAL_VERSION]: [
-		PermissionGrantEntityType.PROPOSAL_VERSION,
-	],
-	[PermissionGrantEntityType.APPLICATION_FORM]: [
-		PermissionGrantEntityType.APPLICATION_FORM,
-	],
-	[PermissionGrantEntityType.APPLICATION_FORM_FIELD]: [
-		PermissionGrantEntityType.APPLICATION_FORM_FIELD,
-	],
-	[PermissionGrantEntityType.PROPOSAL_FIELD_VALUE]: [
-		PermissionGrantEntityType.PROPOSAL_FIELD_VALUE,
-	],
-	[PermissionGrantEntityType.SOURCE]: [PermissionGrantEntityType.SOURCE],
-	[PermissionGrantEntityType.BULK_UPLOAD]: [
-		PermissionGrantEntityType.BULK_UPLOAD,
-	],
-	[PermissionGrantEntityType.CHANGEMAKER_FIELD_VALUE]: [
-		PermissionGrantEntityType.CHANGEMAKER_FIELD_VALUE,
-	],
-	[PermissionGrantEntityType.ANY]: [],
-} as const satisfies Record<
-	PermissionGrantEntityType,
-	readonly PermissionGrantEntityType[]
->;
-
-const scopeConditions = {
-	[PermissionGrantEntityType.CHANGEMAKER]: [],
-	[PermissionGrantEntityType.FUNDER]: [],
-	[PermissionGrantEntityType.DATA_PROVIDER]: [],
-	[PermissionGrantEntityType.OPPORTUNITY]: [],
-	[PermissionGrantEntityType.PROPOSAL]: [],
-	[PermissionGrantEntityType.PROPOSAL_VERSION]: [],
-	[PermissionGrantEntityType.APPLICATION_FORM]: [],
-	[PermissionGrantEntityType.APPLICATION_FORM_FIELD]: [],
-	[PermissionGrantEntityType.PROPOSAL_FIELD_VALUE]: [
-		{
-			property: 'baseFieldCategory',
-			operator: 'in',
-			value: Object.values(BaseFieldCategory),
-		},
-	],
-	[PermissionGrantEntityType.SOURCE]: [],
-	[PermissionGrantEntityType.BULK_UPLOAD]: [],
-	[PermissionGrantEntityType.CHANGEMAKER_FIELD_VALUE]: [],
-	[PermissionGrantEntityType.ANY]: [],
-} as const satisfies Record<
-	PermissionGrantEntityType,
-	readonly PermissionGrantCondition[]
->;
 
 interface UnkeyedPermissionGrant {
 	readonly id: Id;
@@ -256,29 +93,6 @@ type PermissionGrant = UnkeyedPermissionGrant &
 type WritablePermissionGrant = Writable<PermissionGrant>;
 
 type WritableUnkeyedPermissionGrant = Writable<UnkeyedPermissionGrant>;
-
-const jsonSchemaTypeForPermissionGrantEntityKeyType: Record<
-	PermissionGrantEntityKeyType,
-	string
-> = {
-	[PermissionGrantEntityKeyType.ID]: 'integer',
-	[PermissionGrantEntityKeyType.SHORT_CODE]: 'string',
-};
-
-const permissionGrantConditionSchema = {
-	type: 'object',
-	properties: {
-		property: { type: 'string' },
-		operator: { type: 'string' },
-		value: {
-			type: 'array',
-			items: { type: 'string' },
-			minItems: 1,
-		},
-	},
-	required: ['property', 'operator', 'value'],
-	additionalProperties: false,
-};
 
 const isWritableUnkeyedPermissionGrant =
 	ajv.compile<WritableUnkeyedPermissionGrant>({
@@ -379,11 +193,6 @@ const coreWritableKeys = new Set([
 	'conditions',
 ]);
 
-const isPermissionGrantKeyedContextEntityType = (
-	contextEntityType: PermissionGrantEntityType,
-): contextEntityType is PermissionGrantKeyedContextEntityType =>
-	contextEntityType in contextEntityKeyProperties;
-
 const getAllowedKeys = (
 	contextEntityType: PermissionGrantEntityType,
 	granteeType: PermissionGrantGranteeType,
@@ -440,17 +249,6 @@ const isWritablePermissionGrant: TypeGuardWithAjvErrors<
 	return true;
 };
 
-const getScopesForContextEntityType = (
-	contextEntityType: PermissionGrantEntityType,
-): readonly PermissionGrantEntityType[] => [
-	...contextEntityTypeNativeScopes[contextEntityType],
-	...universalScopes,
-];
-
-const getConditionsForScope = (
-	scopeKey: PermissionGrantEntityType,
-): readonly PermissionGrantCondition[] => scopeConditions[scopeKey];
-
 /**
  * Returns the grantee + scope + verbs portion of a permission grant that
  * gives the auth user `manage` against `any` scope. Spread into a grant
@@ -472,20 +270,10 @@ const getSelfManageGrantFragment = (
 });
 
 export {
-	contextEntityKeyProperties,
-	getConditionsForScope,
-	getScopesForContextEntityType,
 	getSelfManageGrantFragment,
-	isPermissionGrantKeyedContextEntityType,
 	isWritablePermissionGrant,
-	PermissionGrantEntityKeyType,
-	PermissionGrantEntityType,
-	PermissionGrantGranteeType,
 	type PermissionGrant,
-	type PermissionGrantCondition,
 	type PermissionGrantContextEntity,
-	type PermissionGrantEntityKeyValueType,
-	type PermissionGrantKeyedContextEntityType,
 	type WritablePermissionGrant,
 	type WritableUnkeyedPermissionGrant,
 };
