+++ src/types/index.ts
@@ -35,6 +35,8 @@ export * from './PaginationParameters';
 export * from './PaginationParametersQuery';
 export * from './Permission';
 export * from './PermissionGrant';
+export * from './PermissionGrantCondition';
+export * from './PermissionGrantEntityType';
 export * from './PermissionGrantGranteeType';
 export * from './PermissionGrantVerb';
 export * from './PlatformProviderResponse';
